requirements.txt
transformers
tqdm
pandas
requests
flask
torch
pyPDF2
pdfminer.six
werkzeug